# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Murojaat turkumlash va yo'naltirish tizimi API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Request {
  id: number;
  fullName: string;
  /** @nullable */
  contactInfo?: string | null;
  text: string;
  /** Yo'naltirilgan bo'lim: hr, it, finance, legal, complaints, general */
  department: string;
  /** Bo'lim nomi o'zbek tilida */
  departmentLabel?: string;
  /** AI ishonch darajasi 0-1 */
  confidence: number;
  /**
     * AI turkumlash izohi
     * @nullable
     */
  reasoning?: string | null;
  /** low | medium | high */
  priority?: string;
  /** new | in_progress | resolved | closed */
  status: string;
  createdAt: string;
}

export interface RequestInput {
  /** @minLength 2 */
  fullName: string;
  contactInfo?: string;
  /**
     * Murojaat matni
     * @minLength 10
     */
  text: string;
}

export type RequestStatusUpdateStatus = typeof RequestStatusUpdateStatus[keyof typeof RequestStatusUpdateStatus];


export const RequestStatusUpdateStatus = {
  new: 'new',
  in_progress: 'in_progress',
  resolved: 'resolved',
  closed: 'closed',
} as const;

export interface RequestStatusUpdate {
  status: RequestStatusUpdateStatus;
}

export interface DeptCount {
  department: string;
  departmentLabel: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface PriorityCount {
  priority: string;
  count: number;
}

export interface RequestStats {
  total: number;
  byDepartment: DeptCount[];
  byStatus: StatusCount[];
  byPriority: PriorityCount[];
  recentActivity: Request[];
}

export type ListRequestsParams = {
department?: string;
status?: string;
limit?: number;
};