# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Murojaat turkumlash va yo'naltirish tizimi API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: requests
    description: Murojaat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: Barcha murojaatlarni olish
      parameters:
        - name: department
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Murojaatlar ro'yxati
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Request"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Yangi murojaat yuborish va avtomatik turkumlash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInput"
      responses:
        "201":
          description: Murojaat yaratildi va turkumlandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"

  /requests/stats:
    get:
      operationId: getRequestStats
      tags: [requests]
      summary: Dashboard statistikasi
      responses:
        "200":
          description: Statistika ma'lumotlari
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestStats"

  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Bitta murojaatni olish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Murojaat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "404":
          description: Topilmadi

    patch:
      operationId: updateRequestStatus
      tags: [requests]
      summary: Murojaat holatini yangilash
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestStatusUpdate"
      responses:
        "200":
          description: Yangilandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Request:
      type: object
      required:
        - id
        - fullName
        - text
        - department
        - confidence
        - status
        - createdAt
      properties:
        id:
          type: integer
        fullName:
          type: string
        contactInfo:
          type: ["string", "null"]
        text:
          type: string
        department:
          type: string
          description: "Yo'naltirilgan bo'lim: hr, it, finance, legal, complaints, general"
        departmentLabel:
          type: string
          description: "Bo'lim nomi o'zbek tilida"
        confidence:
          type: number
          description: "AI ishonch darajasi 0-1"
        reasoning:
          type: ["string", "null"]
          description: "AI turkumlash izohi"
        priority:
          type: string
          description: "low | medium | high"
        status:
          type: string
          description: "new | in_progress | resolved | closed"
        createdAt:
          type: string
          format: date-time

    RequestInput:
      type: object
      required:
        - fullName
        - text
      properties:
        fullName:
          type: string
          minLength: 2
        contactInfo:
          type: string
        text:
          type: string
          minLength: 10
          description: "Murojaat matni"

    RequestStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [new, in_progress, resolved, closed]

    RequestStats:
      type: object
      required:
        - total
        - byDepartment
        - byStatus
        - byPriority
        - recentActivity
      properties:
        total:
          type: integer
        byDepartment:
          type: array
          items:
            $ref: "#/components/schemas/DeptCount"
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        byPriority:
          type: array
          items:
            $ref: "#/components/schemas/PriorityCount"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Request"

    DeptCount:
      type: object
      required: [department, departmentLabel, count]
      properties:
        department:
          type: string
        departmentLabel:
          type: string
        count:
          type: integer

    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer

    PriorityCount:
      type: object
      required: [priority, count]
      properties:
        priority:
          type: string
        count:
          type: integer