# lib/api-zod/src/generated/types/deptCount.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Murojaat turkumlash va yo'naltirish tizimi API
 * OpenAPI spec version: 0.1.0
 */

export interface DeptCount {
  department: string;
  departmentLabel: string;
  count: number;
}